example/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
django==3.1.2             # via -r requirements.in
pytz==2020.1              # via django
sqlparse==0.4.1           # via django
factory_boy==2.12.0
